requirements.txt
python-dotenv
newsapi-python
streamlit
schedule
